Incorporate NYC Open Data for borough boundaries.
Found NYC Open Data page with links to GeoJSON data that precisely defines the boundaries for each of the boroughs.  Use this data to render the Choose Location map (the data will eventually be used to ensure we only get StreetView Panorama's from within city limits.  Remove the 'hole' from the Choose Location Map.  Add an event listener to change the color of a borough on mouseover.